fix silly bug in the devices page

+++ resources/migrations/20181007174942-add-metrics-table.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE metrics(
   device_registry_id	      VARCHAR(50) NOT NULL,
-  device_id	                INTEGER     REFERENCES devices(id),
+  device_id	                BIGINT     REFERENCES devices(id),
   core_temp_max	            NUMERIC,
   core_temp_main	          NUMERIC,
   humidity	                NUMERIC,
